feat: create folder for image automatic

+++ app/Http/Controllers/Dashboard/DashboardProjectController.php
@@ -9,11 +9,14 @@ use App\Models\Sektor;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 use App\Exports\ProjectExport;
+use Illuminate\Support\Facades\DB;
 use Maatwebsite\Excel\Facades\Excel;
 
 
 class DashboardProjectController extends Controller
 {
+    private $projectImageFolder = 'project_images';
+
     public function index()
     {
         $projects = Project::with('sektor')->get();
@@ -38,17 +41,21 @@ class DashboardProjectController extends Controller
             'sektor_id' => 'required|exists:sektors.id',
         ]);
 
-        $imagePath = $request->file('image')->store('image_project', 'public');
+        DB::transaction(function () use ($request) {
+            $this->ensureProjectImageFolderExists();
 
-        Project::create([
-            'title' => $request->title,
-            'description' => $request->description,
-            'image' => $imagePath,
-            'lokasi_kecamatan' => $request->lokasi_kecamatan,
-            'tanggal_awal' => $request->tanggal_awal,
-            'tanggal_akhir' => $request->tanggal_akhir,
-            'sektor_id' => $request->sektor_id,
-        ]);
+            $imagePath = $request->file('image')->store($this->projectImageFolder, 'public');
+
+            Project::create([
+                'title' => $request->title,
+                'description' => $request->description,
+                'image' => $imagePath,
+                'lokasi_kecamatan' => $request->lokasi_kecamatan,
+                'tanggal_awal' => $request->tanggal_awal,
+                'tanggal_akhir' => $request->tanggal_akhir,
+                'sektor_id' => $request->sektor_id,
+            ]);
+        });
 
         return redirect()->route('dashboard.project.index')->with('success', 'Project Berhasil Dibuat.');
     }
@@ -82,17 +89,22 @@ class DashboardProjectController extends Controller
             'tanggal_akhir' => 'required|date|after:tanggal_awal',
             'sektor_id' => 'required|exists:sektors.id',
         ]);
+        DB::transaction(function () use ($request, $project) {
+            $data = $request->except('image');
 
-        $data = $request->except('image');
+            if ($request->hasFile('image')) {
+                $this->ensureProjectImageFolderExists();
 
-        if ($request->hasFile('image')) {
-            Storage::disk('public')->delete($project->image);
-            $imagePath = $request->file('image')->store('project_images', 'public');
-            $data['image'] = $imagePath;
-        }
+                if ($project->image) {
+                    Storage::disk('public')->delete($project->image);
+                }
 
-        $project->update($data);
+                $imagePath = $request->file('image')->store($this->projectImageFolder, 'public');
+                $data['image'] = $imagePath;
+            }
 
+            $project->update($data);
+        });
         return redirect()->route('dashboard.project.index')->with('success', 'Project Berhasil Diperbaharui.');
     }
 
@@ -117,5 +129,12 @@ class DashboardProjectController extends Controller
         $project->save();
 
         return redirect()->back()->with('success', 'Status Project Berhasil Diupdate.');
+    }
+
+    private function ensureProjectImageFolderExists()
+    {
+        if (!Storage::disk('public')->exists($this->projectImageFolder)) {
+            Storage::disk('public')->makeDirectory($this->projectImageFolder);
+        }
     }
 }
